target layout for generated io monitor
